refactor code to address requested PR change
- removed magic strings
- created exception.py file for error handlers
- removed redundant code
- created .env for environment variables
- created utils package for test helper functions
- included more accessions
