+++ project-1/exceptions.py
@@ -1,0 +1,19 @@
+# error handlers
+class DataPopulationError(Exception):
+    pass
+
+
+class NewSheetError(Exception):
+    pass
+
+
+class NewSpreadsheetError(Exception):
+    pass
+
+
+class DownloadDataError(Exception):
+    pass
+
+
+class PreprocessError(Exception):
+    pass
